Rename fetchData to fetchProfileData
