+++ client/src/components/Form/ProfileForm.js
@@ -1,15 +1,13 @@
 import React from 'react';
-
-import fetchData from '../../utils/fetchData';
-import SearchBar from '../common/SearchBar';
 import useDropdown from '../../hooks/useDropdown';
 import useLocalStorage from '../../hooks/useLocalStorage';
+import fetchProfileData from '../../utils/fetchProfileData';
+import SearchBar from '../common/SearchBar';
 
 const ProfileForm = ({ setData, setIsError, setIsLoading, navigate }) => {
   const options = ['login', 'extensions/user_id'];
   const [option, Dropdown] = useDropdown('login', 'login', options);
   const [query, setQuery] = useLocalStorage('queryStorage', '');
-
   const URL = process.env.REACT_APP_URL || '';
 
   return (
@@ -17,14 +15,13 @@ const ProfileForm = ({ setData, setIsError, setIsLoading, navigate }) => {
       className="twitch-form"
       onSubmit={e => {
         e.preventDefault();
-        fetchData(
+        fetchProfileData(
           `${URL}/users/${option}=${query}`,
           setData,
           setIsError,
           setIsLoading
         );
 
-        // Access navigate from props - github.com/reach/router/issues/225
         navigate(`${option}/${query}`);
       }}
     >
