+++ client/src/components/results/ProfileResults.js
@@ -1,12 +1,11 @@
 import React from 'react';
-
 import ExtensionsData from '../data/ExtensionsData';
 import ProfileData from '../data/ProfileData';
 
 const ProfileResults = ({ data }) => {
   if (data.length > 0) data = JSON.parse(data);
 
-  let toggle = data.panel;
+  const toggle = data.panel;
 
   return (
     <>
