+++ client/src/index.js
@@ -1,7 +1,6 @@
+import { CSSReset, ThemeProvider } from '@chakra-ui/core';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { ThemeProvider, CSSReset } from '@chakra-ui/core';
-
 import App from './App';
 
 ReactDOM.render(
