+++ client/src/pages/Home.js
@@ -1,10 +1,9 @@
 import React from 'react';
 
-const Home = props => {
+const Home = () => {
   return (
     <div className="Home">
       <h1>Twitch Analytics</h1>
-
       <p>Enter your Twitch Username to see your profile data</p>
     </div>
   );
